refactor: update imports for quarkus-ledger-api module split
Move ActorType, LedgerEntryType from runtime.model to api.model and
LedgerTraceIdProvider from runtime.service to api.spi.

Refs casehubio/quarkus-ledger#73
